refactor: rename classes to enhance modularity
WinningNumbers -> MainNumbers
Numbers -> WinningNumbers
